chore(wizard-refactor): extract _shared helpers (validUuid + welcome-phases + applyStudentExperienceConfig)
Moves three pure helpers out of the 2900-line wizard-tool-executor monolith
into wizard-tool-executor/_shared/. No behaviour change; monolith imports
them and re-exports applyStudentExperienceConfig for backwards-compat with
external callers (chat/route.ts, wizard tests).

Pinned by tests/lib/chat/wizard-tool-executor-dispatcher.test.ts (11 cases)
+ tests/eslint-rules/no-ai-fanout-all.test.ts (17 cases): 28/28 green.
