+++ apps/admin/lib/chat/wizard-tool-executor.ts
@@ -11,135 +11,10 @@
 
 import { updatePlaybookConfig } from "@/lib/playbook/update-playbook-config";
 import { updateDomainConfig } from "@/lib/domain/update-domain-config";
+import { validUuid } from "./wizard-tool-executor/_shared/valid-uuid";
+import { applyStudentExperienceConfig } from "./wizard-tool-executor/_shared/apply-student-experience";
 
-// ── Helpers ─────────────────────────────────────────────
-
-// Canonical wizard phase-id mapping. Used to build sessionFlow.onboarding.phases
-// from the four welcome booleans (#383). Phase ids match the existing
-// sessionFlow.intake key names for consistency. The resolver at
-// session-flow/resolver.ts::resolveOnboarding reads sessionFlow.onboarding
-// first (Priority 1); pedagogy.ts::computeSessionPedagogy reads the legacy
-// Playbook.config.onboardingFlowPhases (Priority 2 in both paths). The wizard
-// mirrors both so it wins regardless of which resolver path runs.
-const WELCOME_PHASE_DEFINITIONS: Array<{
-  key: "welcomeAiIntro" | "welcomeGoals" | "welcomeAboutYou" | "welcomeKnowledgeCheck";
-  phase: string;
-  duration: string;
-  goals: string[];
-}> = [
-  { key: "welcomeAiIntro",        phase: "aiIntro",        duration: "1-2 min", goals: ["Tutor introduces itself and the session frame"] },
-  { key: "welcomeGoals",          phase: "goals",          duration: "2-3 min", goals: ["Share the course goals with the learner"] },
-  { key: "welcomeAboutYou",       phase: "aboutYou",       duration: "2-3 min", goals: ["Discover learner background, motivation, and context"] },
-  { key: "welcomeKnowledgeCheck", phase: "knowledgeCheck", duration: "3-5 min", goals: ["Brief diagnostic to gauge starting level"] },
-];
-
-/**
- * Build the student-experience portion of a Playbook.config from wizard
- * setupData. Writes BOTH the legacy shape (`welcome` / `nps` / `surveys`)
- * AND the new canonical `sessionFlow.intake` shape (#216 mirror pattern).
- *
- * Mutates `configUpdate` in place. Skips fields already present (idempotent).
- *
- * Bag keys (stable contract):
- *   welcomeGoals / welcomeAboutYou / welcomeKnowledgeCheck / welcomeAiIntro: boolean
- *   welcomeKnowledgeCheckMode: "mcq" | "socratic"  (NEW for #222)
- *   npsEnabled: boolean
- *
- * @param setupData wizard flow-bag (record of values keyed by bag key)
- * @param configUpdate mutable Playbook.config working copy
- * @param contextLabel string used in observability warnings ("create_course (existing)" etc.)
- * @param entityId    playbookId / courseId for observability
- */
-export function applyStudentExperienceConfig(
-  setupData: Record<string, unknown> | undefined,
-  configUpdate: Record<string, unknown>,
-  contextLabel: string,
-  entityId: string,
-): void {
-  // ── Welcome (legacy) — always written by wizard until Phase 5 cleanup. ──
-  if (!configUpdate.welcome) {
-    const welcomeKeysSet = ["welcomeGoals", "welcomeAboutYou", "welcomeKnowledgeCheck", "welcomeAiIntro"]
-      .filter((k) => setupData?.[k] !== undefined).length;
-    if (welcomeKeysSet === 0) {
-      console.warn(
-        `[wizard-tool-executor] ${contextLabel} called without explicit welcome flags — falling back to DEFAULT_WELCOME_CONFIG. id=${entityId}`,
-      );
-    }
-    configUpdate.welcome = {
-      goals: { enabled: setupData?.welcomeGoals !== false },
-      aboutYou: { enabled: setupData?.welcomeAboutYou !== false },
-      knowledgeCheck: { enabled: setupData?.welcomeKnowledgeCheck === true },
-      aiIntroCall: { enabled: setupData?.welcomeAiIntro === true },
-    };
-  }
-
-  // ── sessionFlow.intake (new shape) — mirror of welcome plus deliveryMode. ──
-  // Always set so the resolver / editor see the same source of truth. The
-  // mirror layer is removed in Phase 5 (#220) once legacy fields are dropped.
-  const existingSessionFlow = (configUpdate.sessionFlow as Record<string, unknown> | undefined) ?? {};
-  const welcome = configUpdate.welcome as {
-    goals: { enabled: boolean };
-    aboutYou: { enabled: boolean };
-    knowledgeCheck: { enabled: boolean };
-    aiIntroCall: { enabled: boolean };
-  };
-  const deliveryMode: "mcq" | "socratic" =
-    setupData?.welcomeKnowledgeCheckMode === "socratic" ? "socratic" : "mcq";
-
-  // ── sessionFlow.onboarding.phases + onboardingFlowPhases mirror (#383). ──
-  // The first-call structural template resolver reads sessionFlow.onboarding
-  // first (new path) or Playbook.config.onboardingFlowPhases (old path); both
-  // must be written so the educator's welcome-phase choice actually fires at
-  // runtime. Empty array when all four are off — the resolver's truthy-object
-  // check still wins Priority 1, so we don't fall through to INIT-001.
-  const welcomeEnabledMap: Record<string, boolean> = {
-    welcomeGoals: welcome.goals.enabled,
-    welcomeAboutYou: welcome.aboutYou.enabled,
-    welcomeKnowledgeCheck: welcome.knowledgeCheck.enabled,
-    welcomeAiIntro: welcome.aiIntroCall.enabled,
-  };
-  const onboardingPhases = WELCOME_PHASE_DEFINITIONS
-    .filter((def) => welcomeEnabledMap[def.key])
-    .map((def) => ({ phase: def.phase, duration: def.duration, goals: [...def.goals] }));
-
-  configUpdate.sessionFlow = {
-    ...existingSessionFlow,
-    intake: {
-      goals: { enabled: welcome.goals.enabled },
-      aboutYou: { enabled: welcome.aboutYou.enabled },
-      knowledgeCheck: { enabled: welcome.knowledgeCheck.enabled, deliveryMode },
-      aiIntroCall: { enabled: welcome.aiIntroCall.enabled },
-    },
-    onboarding: { phases: onboardingPhases },
-  };
-  configUpdate.onboardingFlowPhases = { phases: onboardingPhases };
-
-  // ── NPS — top-level config field, mirrored to surveys.post.enabled for
-  // structured-mode rail compatibility (existing pattern). ──
-  if (!configUpdate.nps) {
-    configUpdate.nps = {
-      enabled: setupData?.npsEnabled !== false,
-      trigger: "mastery" as const,
-      threshold: 80,
-    };
-  }
-  if (!configUpdate.surveys) {
-    const nps = configUpdate.nps as { enabled: boolean };
-    configUpdate.surveys = {
-      post: { enabled: nps.enabled },
-    };
-  }
-}
-
-/** Return the string only if it looks like a real UUID (v4). Rejects slugs, made-up prefixed IDs, etc. */
-function validUuid(value: unknown): string | undefined {
-  if (typeof value !== "string") return undefined;
-  // Standard UUID v4 pattern — also accepts Prisma cuid/cuid2 (25+ alphanum chars)
-  if (/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i.test(value)) return value;
-  // Prisma CUID (starts with c, 25 chars) or CUID2 (24+ chars alphanumeric)
-  if (/^c[a-z0-9]{24,}$/i.test(value)) return value;
-  return undefined;
-}
+export { applyStudentExperienceConfig };
 
 /** Resolve existing institution by name, or create institution + domain + link user. */
 async function ensureInstitutionAndDomain(
